requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Redis
redis==5.0.1
aioredis==2.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0

# Message Queue
aio-pika==9.4.0
celery==5.3.4

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
sentry-sdk==1.40.0

# Utilities
python-dotenv==1.0.0
pytz==2023.3
uuid==1.30

# Payment Processing
stripe==7.4.0

# Email & SMS
sendgrid==6.11.0
twilio==8.10.0

# QR Code & PDF Generation
qrcode[pil]==7.4.2
reportlab==4.0.7
Pillow==10.1.0

# WebSocket
websockets==12.0
python-socketio==5.10.0

# OAuth2 & 2FA
authlib==1.2.1
pyotp==2.9.0

# Analytics & Data Processing
pandas==2.1.3
numpy==1.26.2
xlsxwriter==3.1.9

# Templates
Jinja2==3.1.2

# Rate Limiting
slowapi==0.1.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
factory-boy==3.3.0
faker==22.1.0
httpx==0.26.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0